-- 07.08.2020 ----

+++ test_l.py
@@ -17,7 +17,7 @@ with open(filename, newline='') as f:
     i=0
     for row in rows:
 
-        x = tf.constant(list(np.float_(row[0:9])), dtype='float64')
+        x = tf.constant(list(np.float_(row[0:9])), dtype='float32')
         my_sum = c.ShaperLayer(2)
         y = my_sum(x)
         if(y > 0):
